+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 
 ### Changed
 
+- Performance and logging tuning
 - Configuration is now stored in /etc/microlab instead of backend/config.py (#141).
 - Recipe data was moved from /backend/recipes/files to being stored by default at /var/lib/microlab/recipes. Recipes in the repository are kept at backend/data/recipes and copied to their proper storage location at launch. (#141).
 - Hardware configuration files were moved from backend/hardware to being stored by default at /var/lib/microlab/controllerhardware and /var/lib/microlab/labhardware. Hardware configurations in the repository are kept at backend/data/hardware and copied to their proper storage location at launch. (#141).